requirements.txt
minecraft-launcher-lib
requests
pillow
plyer
pystray